test(hir): relocate HIR span recovery helpers tests to src/hir/tests.rs
Moved all tests for HIR span recovery helpers and macro-expansion behavior from src/hir/mod.rs into a new dedicated test module src/hir/tests.rs. This improves code organization and readability by separating tests from implementation.
